Rework with new trait (#28)
This is a fairly large PR, but in broad strokes, it:

* Introduces the trait described by #25
* Implements the behavior and "data format" described by #25
    * One note: we are now able to store the whole path as the key, instead of just a hash
* Implements a non-sequential-storage based test impl of the trait, allowing for more direct testing

The trait lifetimes got ROUGH, and due to that, I ended up needing to compromise on some interface niceness, but I was able to make it okay-ish with some macros. End-users won't see this grossness, only folks that try to implement the trait, and if they use the blanket sequential-storage impl: they won't ever need to touch it.

Also as a note, I had hoped at least the test impl of the trait would be `Send`, but there are still issues (cc https://github.com/rust-lang/rust/issues/100013), so I've switched to just using `LocalSet` for all testing to not require `Send` in testing.